Sembla el css i js a index tira bastant, hi ha el TAR i el FB

+++ jocs/forms.py
@@ -4,7 +4,7 @@ from django import forms
 class JocForm (forms.Form):
     imatge = forms.ImageField(label='Selecciona un archiu')
     nom = forms.CharField(max_length=120, label = 'nom')
-    preuE=forms.IntegerField(label="preuE")
+    preuE=forms.DecimalField(max_digits=7,decimal_places=2,label="preuE")
     preuG=forms.IntegerField(label="preuG")
     descripcio= forms.CharField(label="descripcio",widget=forms.Textarea)
